+++ apps/dev/app/api/cms/route.ts
@@ -1,0 +1,7 @@
+import {createHandler} from 'alinea/next'
+import {cms} from '@/cms'
+
+const handler = createHandler(cms)
+
+export const GET = handler
+export const POST = handler