+++ src/core/util/Slugs.ts
@@ -2,8 +2,11 @@
 // A special case is made for punctuation which can be included in emojis,
 // which in this case would be followed by the emoji variation selector
 // For example we'll strip # but not #️⃣
-const strip = /([^\p{L}\p{N}\p{Emoji}\p{Emoji_Component}]|\p{P}(?!\u{fe0f}))+/gu
-const ignore = /[\'’]/g
+const strip =
+  /([^\p{L}\p{N}\p{M}\p{Emoji}\p{Emoji_Component}]|\p{P}(?!\u{fe0f}))+/u
+const stripGlobal = new RegExp(strip.source, 'gu')
+const ignore = /['’]/g
+const stripAccents = /(\p{Script=Latin})\p{M}+/gu
 const separator = '-'
 const trim = new RegExp(`^${separator}+|${separator}+$`, 'g')
 
@@ -12,9 +15,13 @@ export function isSeparator(char: string): boolean {
 }
 
 export function slugify(input: string): string {
-  return input
-    .replace(ignore, '')
-    .replace(strip, separator)
-    .replace(trim, '')
-    .toLowerCase()
+  return (
+    input
+      .normalize('NFKD')
+      .replace(stripAccents, '$1')
+      .replace(ignore, '')
+      .replace(stripGlobal, separator)
+      .replace(trim, '')
+      .toLowerCase()
+  )
 }